Refactor database pages common code
Extracts db_common.inc.php code to the Common::database() method and
removes the include file.
