+++ libraries/classes/Controllers/Database/TriggersController.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\Controllers\Database;
 
+use PhpMyAdmin\Common;
 use PhpMyAdmin\Rte\Triggers;
 use PhpMyAdmin\Url;
 use PhpMyAdmin\Util;
@@ -38,7 +39,7 @@ class TriggersController extends AbstractController
                 include_once ROOT_PATH . 'libraries/tbl_common.inc.php';
             } else {
                 $table = '';
-                include_once ROOT_PATH . 'libraries/db_common.inc.php';
+                Common::database();
 
                 list(
                     $tables,
